requirements.txt
et-xmlfile==1.0.1
IMAPClient==2.1.0
ip==19.2.3
jdcal==1.4.1
lxml==4.5.0
openpyxl==3.0.3
p==19.2.3
Pillow==7.1.1
PyPDF2==1.26.0
python-docx==0.8.10
selenium==3.141.0
six==1.14.0
style==1.1.0
update==0.0.1
urllib3==1.25.8
